requirements.txt
alembic==1.14.1
bcrypt==4.1.3
bleach==6.2.0
blinker==1.9.0
chardet==5.2.0
click==8.1.8
dnspython==2.7.0
email-validator==2.0.0
Flask==3.0.2
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
pillow==11.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.1.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.1.2
